Fix Docker release script (copying artifacts to host)
Use `docker cp <container:path> <host path>`, which creates the files under
the host user context and not the container's root user. It doesn't support
wildcards unfortunately.
The release and source archives are copied to the host's current directory
after the container has run but before removing it.

Also don't hardcode LDC_VERSION and -_SUFFIX in the script, but require them
as host environment variables and forward them to the container. All other
environment variables (BUILD_WITH_LTO, DUB_VERSION, ...) will be set to their
default values in the container.